feat/#33: interface 업데이트
