+++ types/category/client.ts
@@ -1,9 +1,6 @@
-export interface MyPageUserInformation {
-  name: string;
-  email: string;
-}
-
-export interface MyPageHistory {
-  recCount: number;
-  recData: { id: number; userId: number; url: string; recommendSize: string }[];
+export interface AllCategory {
+  id: string;
+  categoryName: string;
+  isPin: boolean;
+  productNum: number;
 }
