+++ types/common.ts
@@ -1,21 +1,9 @@
-export interface closetData {
-  id: string;
-  image?: string;
-  productName: string;
-  size?: string;
-  memo?: string;
-  mallName?: string;
-  isRecommend?: boolean;
-  isPin: boolean;
-  link?: string;
-}
-
 export interface ThumbNailData {
   id: string;
   image?: string;
   productName?: string;
   size?: string;
-  memo?: string;
+  memo?: string | null;
   isRecommend?: boolean;
   isPin: boolean;
 }
