+++ types/mypage/client.ts
@@ -1,0 +1,9 @@
+export interface MyPageUserInformation {
+  name: string;
+  email: string;
+}
+
+export interface MyPageHistory {
+  recCount: number;
+  recData: { id: number; userId: number; url: string; recommendSize: string }[];
+}
