+++ types/mypage/remote.ts
